// packages/react/src/components/RichText/internal/Extensions/Mention/types.ts
export interface MentionNodeAttrs {
  id: string
  label: string
  image_url?: string
  href?: string
}

export interface MentionListRef {
  onKeyDown: (props: { event: KeyboardEvent }) => boolean
}

export interface MentionItemComponentProps {
  item: MentionedUser
  index: number
  selected: boolean
}

export type MentionedUser = {
  id: string | number
  label: string
  image_url?: string
  href?: string
}

export type MentionsConfig = {
  onMentionQueryStringChanged?: (
    queryString: string
  ) => Promise<MentionedUser[]> | undefined
  users: MentionedUser[]
}